Following Ben's advice - it turns out the Raw Nimrod Data is reported in mm/hr * 32. I have therefore divided values by 32.
Further - I noticed that I was summing all columns during the resample
step at the end which is not correct. Have now corrected to sum the
mean rainfall, total ranfall. then max of the max rain and std. then min
of the min rainfall.

Last thing was to report the max, min and std as mm/hr and I also
dropped the total rainfall column.